Simplify time and duration logic

+++ storage_fs.go
@@ -162,10 +162,9 @@ func (s *FileStore) Recover(path string, fileInfo os.FileInfo, err error) error 
 	}
 	modTime := fileInfo.ModTime()
 	deathTime := modTime.Add(lifeTime)
-	if lifeTime > 0 {
-		if deathTime.Before(startTime) {
-			return os.Remove(path)
-		}
+	lifeLeft := deathTime.Sub(startTime)
+	if lifeTime > 0 && lifeLeft <= 0 {
+		return os.Remove(path)
 	}
 	size := fileInfo.Size()
 	s.Lock()
@@ -174,7 +173,6 @@ func (s *FileStore) Recover(path string, fileInfo os.FileInfo, err error) error 
 		return ErrReachedMax
 	}
 	s.stats.makeSpaceFor(size)
-	lifeLeft := deathTime.Sub(startTime)
 	cached := fileCache{
 		header: genHeader(id, modTime, size),
 		path:   path,
